Fix failing Python 3.5 tests
Order of keys or values returned from an iteration on a dictionary
is not guarenteed. Sorting on attributes is required to ensure that
tests pass across versions of Python.

_Note_: This `compare` function is needlessly complicated.
The use of `try` and `except` catches too many cases and these tests
will fail to catch bugs introduced in future development.
Additionally, makes debugging very challenging. This is in severe need for a
refactor. This can be written more cleanly using `if` and `else`.