DONE: In parse() the base folder must be provided for any include files DONE: In parse(), need to strip off ';' when parsing location for include files DONE: In parse(), sometimes schedules can be on the same row. Considered this case DONE: auction, eventgen, metrics, power_metrics, voltdump, currdump, jsondump triplex_meter, passive_controller, stub_bidder, evcharger_det, diesel_dg, battery, solar, group_recorder, multi_recorder, collector and assert objects weren't supported in the gridlab-d schema. collector was added to the schema and the rest are ignored DONE: The raise_taps attribute of a regulator should be an integer but it's a float. Parsed to a float and then cast to int. DONE: The tap_position attribute of a capacitor should be an integer but it's a float. Parsed to a float and then cast to int. DONE: Sub-objects in gridlab-d were causing problems for our parser. Now any sub-object is ingored with a warning message (this was partially implemented before). e.g. in the code below recorder is a sub-object which we don't want:
object regulator:3920 {
     name R5-25-00-1_reg_1;
     from R5-25-00-1_node_953;
     to R5-25-00-1_meter_15;
     phases ABCN;
     configuration regulator_configuration:3921;
     object recorder {
          property tap_A,tap_B,tap_C,power_in_A.real,power_in_A.imag,power_in_B.real,power_in_B.imag,power_in_C.real,power_in_C.imag,power_in.real,power_in.imag;
          limit 1440;
          interval 60;
          file "R5-25.00-1_reg_output.csv";
     };
}

DONE: There was a strange bug where self.all_gld_objects would be persistent accross multiple instances when run sequentially. This is reset in the parse function to address this (not sure about the root cause though?)
DONE: some of the impdance matricies were in string form and needed to be converted to complex numbers in the reader
DONE: powerN_rating wasn't included in the schema (even though it's not used). Added to avoid parsing errors
DONE: phase_loss_protection wasn't included in the schema (even though it's not used). Added to avoid parsing errors
DONE: Added check of what level of nesting we are in for objects so that we ignore all levels of nesting (not just first)

TEMPORARY: wintrub_dg isn't represnted in our schema (and I couldn't find info about it online?) Currently ignoring

TEMPORARY: Reconfigured the kron reduction system to use the kron reduction from the abstract reader. Updated the code to be better. However, serious code review needed to determine the correct impedance matrix to hold, and review how the kron reductions are done in the gridlabd reader and in general (including with underground lines)

NOTE: The following line spacing is provided. Is this a problem? Can't handle this geometry for underground lines....

object line_spacing {
        distance_AN 0.000000;
        name 57825line_configuration40901line_spacing40902;
        distance_CN 0.000000;
        distance_BC 0.500000;
        distance_AB 0.500000;
        distance_BN 0.000000;
        distance_AC 1.000000;
};

object line_spacing {
        distance_AN 0.0;
        name 825117-LINESPACING;
        distance_CN 0.0;
        distance_BC 0.0;
        distance_AB 0.0;
        distance_AC 0.0;
        distance_BN 0.0;
};