cyme reader handle multiple networks
a cyme "network" is either a substation or a feeder. each network has a head node. the old behavior was to set all sources (i.e. head nodes of networks) as PowerSource.is_sourcebus. With multiple sources DiTTo raises a value error (multiple sources in general should be allowed, but for the openDSS writer the is_sourcebus field is used to define the network / Vsource). So maybe need a clearer definition of what is a DiTTo PowerSource and why we should always have one (slack bus?)